requirements.txt
google-cloud-aiplatform
pandas
scikit-learn
